update: move parser finding func, component imports
